Add initial disk cache implementation
