+++ .pre-commit-config.yaml
@@ -3,3 +3,8 @@ repos:
     rev: 23.12.0
     hooks:
       - id: black
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.1.7
+    hooks:
+      - id: ruff
+        args: [ --fix ]
