+++ dqp/storage.py
@@ -1,5 +1,4 @@
 import os
-from typing import Dict
 
 import msgpack
