+++ pyproject.toml
@@ -22,7 +22,6 @@ ruff = "^0.1.7"
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
-
 [tool.ruff]
 extend-select = ["I"]
